requirements.txt
faiss-cpu>=1.7.2
numpy>=1.19.5
pyzmq>=22.0.0
msgpack>=1.0.2
